# docs/swagger.yaml
basePath: /
definitions:
...
  models.DeleteUserResponse:
    properties:
      message:
        type: string
    type: object
  models.ErrorResponse:
    properties:
      code:
        example: INVALID_TABLE
        type: string
      details:
        example: Table 'invalid_table' does not exist or is not accessible
        type: string
      error:
        example: Invalid table name
        type: string
    type: object
...
  models.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
      tenant_slug:
        description: 'Optional: specify tenant for login'
        type: string
    required:
    - email
    - password
    type: object
  models.LoginResponse:
    properties:
      tenant_id:
        type: string
      tenant_slug:
        type: string
      token:
        type: string
      user:
        $ref: '#/definitions/models.User'
    type: object
  models.SignUpRequest:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        minLength: 8
        type: string
      tenant_slug:
        description: 'Optional: create/join tenant during signup'
        type: string
    required:
    - email
    - first_name
    - last_name
    - password
    type: object
  models.SignUpResponse:
    properties:
      message:
        type: string
      user:
        $ref: '#/definitions/models.User'
    type: object
  models.SwitchTenantRequest:
    properties:
      tenant_id:
        type: string
    required:
    - tenant_id
    type: object
  models.SwitchTenantResponse:
    properties:
      message:
        type: string
      tenant_id:
        type: string
      tenant_slug:
        type: string
      token:
        type: string
    type: object
  models.Tenant:
    properties:
      created_at:
        type: string
      domain:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      slug:
        type: string
      updated_at:
        type: string
    type: object
...
  models.UpdateUserRequest:
    properties:
      first_name:
        type: string
      is_active:
        type: boolean
      last_name:
        type: string
    type: object
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      last_name:
        type: string
      updated_at:
        type: string
    type: object
...
info:
  contact: {}
  description: Directus-style API with Role-Based Access Control (RBAC). A powerful,
    generic API that provides CRUD operations for any database table with comprehensive
    security, multi-tenancy, and dynamic schema management.
  title: Basin API
  version: 1.0.0
paths:
  /auth/context:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get Auth Context
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LoginResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login
      tags:
      - auth
  /auth/me:
    get:
      description: Retrieve information about the currently authenticated user. Requires
        valid JWT Bearer token or API key.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Get current user
      tags:
      - auth
  /auth/signup:
    post:
      consumes:
      - application/json
      parameters:
      - description: Sign up payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.SignUpRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.SignUpResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Sign Up
      tags:
      - auth
  /auth/switch-tenant:
    post:
      consumes:
      - application/json
      parameters:
      - description: Switch tenant payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.SwitchTenantRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SwitchTenantResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Switch Tenant
      tags:
      - auth
  /auth/tenants:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Tenant'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get User Tenants
      tags:
      - auth
  /auth/users/{id}:
    delete:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DeleteUserResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete User
      tags:
      - auth
    put:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Update payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update User
      tags:
      - auth
...
securityDefinitions:
  ApiKeyAuth:
    description: 'API key for programmatic access (format: Bearer YOUR_API_KEY)'
    in: header
    name: Authorization
    type: apiKey
  BearerAuth:
    description: JWT Bearer token for user authentication
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"